.github/workflows/requirements.txt
# these requirements are for the config.py and submit_spark_job.py that helps package recipe requirements for EMR job runs
s3fs
boto3
tenacity
requests
venvception>=0.0.5
jupyter-repo2docker